@@ -9,9 +9,9 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  casdoor_flutter_sdk: ^1.4.1
+  casdoor_flutter_sdk: ^1.10.0
   cupertino_icons: ^1.0.2
-  shared_preferences: ^2.0.8
+  shared_preferences: ^2.5.3
 
 dev_dependencies:
   flutter_test:
