@@ -5,10 +5,14 @@
 import FlutterMacOS
 import Foundation
 
-import casdoor_flutter_sdk
+import desktop_webview_window
+import flutter_inappwebview_macos
+import path_provider_foundation
 import shared_preferences_foundation
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  CasdoorFlutterSdkPlugin.register(with: registry.registrar(forPlugin: "CasdoorFlutterSdkPlugin"))
+  DesktopWebviewWindowPlugin.register(with: registry.registrar(forPlugin: "DesktopWebviewWindowPlugin"))
+  InAppWebViewFlutterPlugin.register(with: registry.registrar(forPlugin: "InAppWebViewFlutterPlugin"))
+  PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   SharedPreferencesPlugin.register(with: registry.registrar(forPlugin: "SharedPreferencesPlugin"))
 }
